backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
openpyxl
pytest
pytest-asyncio
pytest-cov
